requirements.txt
appdirs
Django
django-cors-headers
django-extensions
django-geojson
django-leaflet
djangorestframework
djangorestframework-gis
Jinja2
MarkupSafe
numpy
oauthlib
packaging
pandas
psycopg2
pyparsing
python-dateutil
pytz
requests
requests-oauthlib
shapely
six
